functions/generate_2fa_v2/requirements.txt
psycopg2-binary==2.9.10
qrcode==8.2
Pillow==11.3.0
cryptography==42.0.5
pyotp==2.9.0 